Add an event to track readers on rendererers
We want to be able to respond to different identifiers
being rendered on the same renderer, so a mechansim
to do so has been added in. E.g. the alignment reader
can watch the renderer to check for anyone rendering
on it, so that it can modified indexes to handle
tweaked sequences